refactor(settings): extract Units, Location, Panels, System into components/
SettingsScreen.kt was accumulating every section inline, making it hard
to navigate. Move the Units, Location, Panels, and System sections into
their own files under ui/settings/components/, each calling the row
primitives extracted in the previous step. Panels also takes its private
visibleCalendarsSummary/MultiSelectRow/MultiSelectDialog helpers with it,
since they are exclusive to that section.

Pure structural move: no behavior change. The Map and Display sections
are left inline for now (later tasks in this restructure extract them).
